Extract methods to files.
